Moved some things out of resolver.py for re-usabilty
This is a fixup of a bad rebase.
