+++ fluent/utils.py
@@ -44,3 +44,16 @@ def partition(iterable, predicate):
             other.append(i)
 
     return matching, other
+
+
+def numeric_to_native(val):
+    """
+    Given a numeric string (as defined by fluent spec),
+    return an int or float
+    """
+    # val matches this EBNF:
+    #  '-'? [0-9]+ ('.' [0-9]+)?
+    if '.' in val:
+        return float(val)
+    else:
+        return int(val)
